fixed warning

+++ helloworld/helloworld.cpp
@@ -31,7 +31,7 @@ void print_values(const Args&... args)
      const Arg argArray[] = {args...};
      printf("size = %ld\n", sizeof ...(Args) );
     // printf((int)argArray.type); 
-    for (int i=0; i  <  sizeof ...(Args) ; i++)
+    for (uint64_t i=0; i  <  sizeof ...(Args) ; i++)
     {
         printf("type = %d\n", (int)argArray[i].type);
     }
